test(agent): every agent package now builds and tests on a non-Linux dev host
internal/webui, internal/spotify and internal/syscheck called Linux-only
syscalls (statfs, SysProcAttr.Setsid, /proc reads) directly in
production code, so the three packages carrying all recent group,
recall and Spotify work could not even compile - let alone run their
tests - on the maintainer's Windows machine. The syscalls now live
behind build-tagged shims: the Linux build is byte-for-byte the same
logic, other hosts fail open (unknown free space never blocks, no
session detach for helpers that never run there). go build ./... and
go test ./... are green on windows/amd64 and the GOARM=5 cross-compile
is unchanged.
